verifying that the number of  users following equals the number of users being added to list

+++ following_migration_twitter_script.py
@@ -1,10 +1,10 @@
 import tweepy
 
-PUBLIC_API_KEY = 'zEAvFiqQI8HhhemPO3aFyNzT2'
-SECRET_API_KEY = 'RFXB7VbbOzNs1nHo1BAEvuZArieqWVzgXtnBwz3lMJob4qraJk'
+PUBLIC_API_KEY = 'rmX8xp2bw9YQujntaIJZjCPVN'
+SECRET_API_KEY = 'hjbzXudPrBHr8wkhpZm3LRPGYcSU5bQjuxyvNhwKLTxct3QN26'
 
-ACCESS_TOKEN_PUBLIC = '1201977948129021953-rXYtx0sRPCirC38khhTq75U6eLyTJN'
-ACCESS_TOKEN_SECRET = '7RolN6uSrgdzltnQEylMfYK6NAAWJHvejPsaELyXG7ll9'
+ACCESS_TOKEN_PUBLIC = '1201977948129021953-xnRqYL8OrTga3x6J0GKsRDaAvhFPbx'
+ACCESS_TOKEN_SECRET = 'HlKNMoupQDR2rxMobzFKbiyiHxgV8tbEF68eMG7zM1u5X'
 
 LIST_NAME = 'Users-Following'
 LIST_PRIVATE = True     #Set to False if you want the list public
@@ -31,9 +31,26 @@ for current_list in api.lists_all():
 if not list_already_exists:
     our_list = api.create_list(LIST_NAME, mode=list_mode, description=LIST_DESCRIPTION)
 
-#Iterate through friends adding them to list then unfollowing them
+#Get total count of those we are following
+following_count = len(api.friends_ids())
+
+num_added_to_list = 0
+#Iterate through friends adding them to list
 for friend_id in api.friends_ids():
     api.add_list_member(list_id=our_list.id, user_id=friend_id, owner_id=api.me().id)
-    api.destroy_friendship(friend_id)
+    num_added_to_list += 1
 
-print('Script has finished running.')
+#Verify that every person we're following is added to list
+if num_added_to_list == following_count:
+    verified = True
+else:
+    verified = False
+
+if verified:
+    #Iterate through friends unfollowing them
+    for friend_id in api.friends_ids():
+        api.destroy_friendship(friend_id)
+    print('Script has finished running successfully.')
+else:
+    print("ERROR - Not all follows were added to list, not unfollowing users.")
+    print('Script has been terminated due to error.')
